sdk: generalize the swap router
